function/requirements.txt
functions-framework==3.8.1
google-cloud-bigquery==3.21.0
google-cloud-secret-manager==2.20.0
google-auth==2.29.0
